refactor: 필터링 dropdown view로 분리
