+++ src/controllers/RestaurantController.js
@@ -2,8 +2,6 @@ import { FORM_INPUT_QUERY, LOCALSTORAGE_KEY } from '../constant/constants';
 import RestaurantService from '../domain/RestaurantService';
 import { $ } from '../utils/querySelector';
 import OutputView from '../views/OutputView';
-import Dropdown from '../components/Common/Dropdown';
-import { FILTER_DROPDOWN_PROPS, SORT_DROPDOWN_PROPS } from '../constant/options';
 
 class RestaurantController {
   #restaurantList;
@@ -34,13 +32,7 @@ class RestaurantController {
   }
 
   showFilterDropdown() {
-    const filterContainer = $('.restaurant-filter-container');
-
-    const filterDropdown = Dropdown(FILTER_DROPDOWN_PROPS);
-    const sortDropdown = Dropdown(SORT_DROPDOWN_PROPS);
-
-    filterContainer.insertAdjacentHTML('beforeend', filterDropdown);
-    filterContainer.insertAdjacentHTML('beforeend', sortDropdown);
+    OutputView.renderFilterDropdown();
   }
 
   updateRestaurantList() {
@@ -54,6 +46,7 @@ class RestaurantController {
 
   showAddRestaurantModal() {
     const addRestaurantButton = $('.gnb__button');
+
     addRestaurantButton.addEventListener('click', () => {
       OutputView.renderAddRestaurant(this.#restaurantList);
       this.manageAddRestaurantFormEvents();
@@ -73,6 +66,7 @@ class RestaurantController {
       alert(isAddedText);
 
       if (!isAdded) return;
+
       this.updateRestaurantList();
       OutputView.closeModal();
     });
@@ -122,7 +116,7 @@ class RestaurantController {
       if (favoriteButton) {
         this.changeFavoriteButton(restaurantId, favoriteButton);
       } else if (restaurantItem) {
-        this.showDetailRestaurantModal(restaurantItem.id);
+        this.showDetailRestaurantModal(restaurantId);
       }
     });
   }
