+++ src/views/OutputView.js
@@ -2,8 +2,20 @@ import AddRestaurantModal from '../components/Modal/AddRestaurantModal';
 import DetailRestaurantModal from '../components/Modal/DetailRestaurantModal';
 import RestaurantItem from '../components/Restaurant/RestaurantItem';
 import { $ } from '../utils/querySelector';
+import Dropdown from '../components/Common/Dropdown';
+import { FILTER_DROPDOWN_PROPS, SORT_DROPDOWN_PROPS } from '../constant/options';
 
 const OutputView = {
+  renderFilterDropdown() {
+    const filterContainer = $('.restaurant-filter-container');
+
+    const filterDropdown = Dropdown(FILTER_DROPDOWN_PROPS);
+    const sortDropdown = Dropdown(SORT_DROPDOWN_PROPS);
+
+    filterContainer.insertAdjacentHTML('beforeend', filterDropdown);
+    filterContainer.insertAdjacentHTML('beforeend', sortDropdown);
+  },
+
   renderRestaurantList(restaurantList) {
     const restaurantListContainer = $('.restaurant-list-container');
